add product controller -> list controller & image picker controller
