+++ lib/controllers/image_picker_controller.dart
@@ -1,0 +1,18 @@
+import 'dart:io';
+
+import 'package:get/get.dart';
+import 'package:image_picker/image_picker.dart';
+
+class ImagePickerController extends GetxController {
+  final ImagePicker picker = ImagePicker();
+
+  Future<File> imgFromCamera() async {
+    final pickedFile = await picker.getImage(source: ImageSource.camera);
+    return File(pickedFile.path);
+  }
+
+  Future<File> imgFromGallery() async {
+    final pickedFile = await picker.getImage(source: ImageSource.gallery);
+    return File(pickedFile.path);
+  }
+}
